Enhancement of CheckLogs
CheckLogs has got new features.
The switch -csav makes the .position files to be stored in the -config directory.  One effect it the the -dir file may be read only.
Amother feature is the "and" feature in the .csv files.  Any of the setup lines may use the & character thus allow for several strings.