V2.8 Fixes and improvments
Added tests (in consoleSts) on state, type and chkday to minimize user errors.
ManFiles. Fixed -pardir to not be zeroed every loop.
ManFiles. Changed sensSTS() to not update Jvackt every loop when an error is found.
monIPAddr. Made the response time calculation better regarding pings.
ServerThread. Put client.close() at the very end to be sure it wont be skipped by an exeption.